lots of curve intersection work and debug
All the functions are in place for intersections between lines,
elliptical arcs and cubic Beziers, including intersections of
offset versions. Think there remains a flaw in the math for the
CC intersection algo, and all the other intersection functions
based on that same algo, and there's cleanup and simplification
to do, but think we've got full intersection coverage for the path
segment types we currently provide. (All but quadratic Bezier.)